move the emoji to unicode function into ui_utils file
